Refactor: Move devtools to view folder
