+++ packages/extension/src/view/devtools/app.css
