+++ packages/extension/src/view/devtools/components/tabs/BounceTrackingTab.tsx
