+++ packages/extension/src/view/devtools/components/tabs/CookieTab/cookieList.tsx
@@ -21,7 +21,7 @@ import React from 'react';
 /**
  * Internal dependencies.
  */
-import { type CookieData } from '../../../../store';
+import { type CookieData } from '../../../../../store';
 import ListItem from './listItem';
 
 interface ICoookieList {
