+++ packages/extension/src/view/devtools/components/tabs/CookieTab/index.tsx
@@ -21,29 +21,17 @@ import React from 'react';
 /**
  * Internal dependencies.
  */
-import { type CookieData } from '../../../../store';
+import { useCookieStore } from '../../../../../app/cookieStore';
+import CookieList from './cookieList';
 
-interface IListItem {
-  cookie: CookieData;
-}
+export const CookieTab = () => {
+  const cookies = useCookieStore(({ state }) => state.cookies);
 
-const ListItem = ({ cookie }: IListItem) => {
   return (
-    <a href="#" className="block hover:bg-secondary">
-      <div className="px-4 py-3 sm:px-6 border-b">
-        <div className="flex items-center justify-between">
-          <div className="font-medium w-full flex items-center justify-between">
-            <span className="text-sm truncate">{cookie.parsedData.name}</span>
-          </div>
-        </div>
-        <div className="mt-0.5">
-          <div className="truncate text-xs text-secondary">
-            <span>{cookie.parsedData.value}</span>
-          </div>
-        </div>
+    <div className="w-full h-full flex flex-col ">
+      <div className="flex-1 overflow-y-scroll ">
+        <CookieList cookies={cookies} />
       </div>
-    </a>
+    </div>
   );
 };
-
-export default ListItem;
