+++ packages/extension/src/view/devtools/components/tabs/CookieTab/listItem.tsx
@@ -21,17 +21,29 @@ import React from 'react';
 /**
  * Internal dependencies.
  */
-import { useCookieStore } from '../../../../app/cookieStore';
-import CookieList from './cookieList';
+import { type CookieData } from '../../../../../store';
 
-export const CookieTab = () => {
-  const cookies = useCookieStore(({ state }) => state.cookies);
+interface IListItem {
+  cookie: CookieData;
+}
 
+const ListItem = ({ cookie }: IListItem) => {
   return (
-    <div className="w-full h-full flex flex-col ">
-      <div className="flex-1 overflow-y-scroll ">
-        <CookieList cookies={cookies} />
+    <a href="#" className="block hover:bg-secondary">
+      <div className="px-4 py-3 sm:px-6 border-b">
+        <div className="flex items-center justify-between">
+          <div className="font-medium w-full flex items-center justify-between">
+            <span className="text-sm truncate">{cookie.parsedData.name}</span>
+          </div>
+        </div>
+        <div className="mt-0.5">
+          <div className="truncate text-xs text-secondary">
+            <span>{cookie.parsedData.value}</span>
+          </div>
+        </div>
       </div>
-    </div>
+    </a>
   );
 };
+
+export default ListItem;
