+++ packages/extension/src/view/devtools/components/tabs/FingerprintingTab.tsx
