+++ packages/extension/src/view/devtools/devtools.html
