+++ packages/extension/src/view/devtools/index.html
